Move gui events from ContainerEventHandler to ClientHandler
